Plist, entitlements, weather + location and more
Adding in injectable weather service
Adding in weather and location service functionality to fetch entire days weather and break it up into parts. This prevents the need to fetch the hourly and the current time. Adding in extensions and helpers to format. Updating the ui components a bit to take in values to show the weather data. Adjusting the ui to show colours based on the weather.

TODOS:
- Unit testing
- Breaking classes / structs into their own files